Not very good updates

+++ src/ghost.h
@@ -55,9 +55,9 @@ class Ghost: public Enemy{
         void LoadAnimation();
 
     private:
-    double x_, y_, angle_;
-    int width_ = 64;
-    int height_ = 64;
+    float x_, y_, angle_;
+    float width_ = 64;
+    float height_ = 64;
     bool exists_ = true;
     int velocity_, x_velocity_, y_velocity_;
     int xOffset_ = 0, yOffset_ = 0, xCollisionDirection_ = 1, yCollisionDirection_ = 1;
